Rename scheduler to petFeedScheduler
