+++ client/android/app/src/main/java/com/sju18001/petmanagement/ui/myPet/MyPetFragment.kt
@@ -1,20 +1,16 @@
 package com.sju18001.petmanagement.ui.myPet
 
-import android.content.Context
-import android.content.Intent
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.lifecycle.ViewModelProvider
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import androidx.viewpager2.widget.ViewPager2
 import com.google.android.material.tabs.TabLayoutMediator
 import com.sju18001.petmanagement.databinding.FragmentMyPetBinding
 import com.sju18001.petmanagement.ui.myPet.petManager.PetManagerFragment
-import com.sju18001.petmanagement.ui.myPet.scheduler.SchedulerFragment
+import com.sju18001.petmanagement.ui.myPet.petFeedScheduler.PetFeedSchedulerFragment
 
 private val TAB_ELEMENTS = listOf("반려동물 관리", "스케쥴러")
 
@@ -56,7 +52,7 @@ class MyPetFragment : Fragment() {
         override fun createFragment(position: Int): Fragment {
             return when(position) {
                 0 -> PetManagerFragment()
-                else -> SchedulerFragment()
+                else -> PetFeedSchedulerFragment()
             }
         }
     }
