+++ client/android/app/src/main/java/com/sju18001/petmanagement/ui/myPet/petFeedScheduler/PetFeedSchedulerFragment.kt
@@ -1,0 +1,21 @@
+package com.sju18001.petmanagement.ui.myPet.petFeedScheduler
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.sju18001.petmanagement.databinding.FragmentPetFeedSchedulerBinding
+
+class PetFeedSchedulerFragment : Fragment() {
+    private var _binding: FragmentPetFeedSchedulerBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = FragmentPetFeedSchedulerBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+}
