+++ client/android/app/src/main/res/layout/fragment_pet_feed_scheduler.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.myPet.scheduler.SchedulerFragment">
+    tools:context=".ui.myPet.petFeedScheduler.PetFeedSchedulerFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
